requirements.txt
asgiref==3.4.1
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==2.0.12
cos-python-sdk-v5==1.9.22
crcmod==1.7
distlib==0.3.6
Django==3.2.17
django-cors-headers==3.10.1
django-filter==21.1
django-redis==5.2.0
django-simpleui==2023.2.10
djangorestframework==3.14.0
drf-extensions==0.7.1
filelock==3.4.1
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
packaging==21.3
platformdirs==2.4.0
prettytable==3.6.0
pycryptodome==3.17
pyecharts==2.0.1
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytz==2022.7.1
QQLoginTool==0.3.0
redis==4.3.5
requests==2.27.1
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
tencentcloud-sdk-python==3.0.829
typing_extensions==4.1.1
urllib3==1.26.14
wcwidth==0.2.6
xmltodict==0.13.0
zipp==3.6.0
